Move mask initializers to utils.initializers
