+++ src/chromatix/elements/phase_masks.py
@@ -4,13 +4,9 @@ from chex import Array, PRNGKey
 from jax.scipy.ndimage import map_coordinates
 from typing import Callable, Optional, Tuple, Union
 from ..field import Field
-from ..functional.phase_masks import (
-    wrap_phase,
-    phase_change,
-    seidel_aberrations,
-    zernike_aberrations,
-)
-from chromatix.elements.utils import register
+from ..functional import wrap_phase, phase_change
+from ..utils import seidel_aberrations, zernike_aberrations
+from .utils import register
 
 __all__ = [
     "PhaseMask",
