+++ src/chromatix/utils/__init__.py
@@ -1,3 +1,4 @@
 from .utils import *
 from .data import *
 from .shapes import *
+from .initializers import *
